+++ jkube-kit/config/resource/src/main/java/org/eclipse/jkube/kit/config/access/ClusterAccess.java
@@ -14,17 +14,17 @@
 package org.eclipse.jkube.kit.config.access;
 
 
-import io.fabric8.kubernetes.client.Config;
-import io.fabric8.kubernetes.client.DefaultKubernetesClient;
-import io.fabric8.kubernetes.client.KubernetesClient;
-import io.fabric8.kubernetes.client.KubernetesClientException;
-import io.fabric8.openshift.client.DefaultOpenShiftClient;
-import io.fabric8.openshift.client.OpenShiftAPIGroups;
-import io.fabric8.openshift.client.OpenShiftClient;
+import java.net.UnknownHostException;
+
 import org.eclipse.jkube.kit.common.KitLogger;
 import org.eclipse.jkube.kit.common.util.OpenshiftHelper;
 
-import java.net.UnknownHostException;
+import io.fabric8.kubernetes.client.Config;
+import io.fabric8.kubernetes.client.KubernetesClient;
+import io.fabric8.kubernetes.client.KubernetesClientBuilder;
+import io.fabric8.kubernetes.client.KubernetesClientException;
+import io.fabric8.openshift.client.OpenShiftAPIGroups;
+import io.fabric8.openshift.client.OpenShiftClient;
 
 /**
  * @author roland
@@ -48,11 +48,11 @@ public class ClusterAccess {
     }
 
     private KubernetesClient createKubernetesClient() {
-        return new DefaultKubernetesClient(createDefaultConfig());
+        return new KubernetesClientBuilder().withConfig(createDefaultConfig()).build();
     }
 
     private OpenShiftClient createOpenShiftClient() {
-        return new DefaultOpenShiftClient(createDefaultConfig());
+        return createKubernetesClient().adapt(OpenShiftClient.class);
     }
 
     private Config createDefaultConfig() {