+++ jkube-kit/config/service/src/test/java/org/eclipse/jkube/kit/config/service/openshift/OpenShiftBuildServiceTest.java
@@ -13,6 +13,7 @@
  */
 package org.eclipse.jkube.kit.config.service.openshift;
 
+import io.fabric8.openshift.client.OpenShiftClient;
 import org.eclipse.jkube.kit.build.service.docker.ArchiveService;
 import org.eclipse.jkube.kit.common.KitLogger;
 import org.eclipse.jkube.kit.common.RegistryConfig;
@@ -23,10 +24,6 @@ import org.eclipse.jkube.kit.config.service.JKubeServiceException;
 import org.eclipse.jkube.kit.config.service.JKubeServiceHub;
 import org.junit.Before;
 import org.junit.Test;
-import org.junit.runner.RunWith;
-import org.mockito.Answers;
-import org.mockito.Mock;
-import org.mockito.junit.MockitoJUnitRunner;
 
 import java.io.IOException;
 import java.util.Collections;
@@ -36,19 +33,17 @@ import static org.junit.Assert.assertThrows;
 import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.Mockito.RETURNS_DEEP_STUBS;
 import static org.mockito.Mockito.mock;
+import static org.mockito.Mockito.spy;
 import static org.mockito.Mockito.times;
 import static org.mockito.Mockito.verify;
 import static org.mockito.Mockito.when;
 
 @SuppressWarnings("unused")
-@RunWith(MockitoJUnitRunner.class)
 public class OpenShiftBuildServiceTest {
 
-  @Mock(answer = Answers.RETURNS_DEEP_STUBS)
   private JKubeServiceHub jKubeServiceHub;
 
-  @Mock(answer = Answers.RETURNS_DEEP_STUBS)
-  private KitLogger mockedKitLogger;
+  private KitLogger kitLogger;
 
   private ImageConfiguration imageConfiguration;
 
@@ -56,6 +51,12 @@ public class OpenShiftBuildServiceTest {
 
   @Before
   public void setUp() {
+    kitLogger = spy(new KitLogger.SilentLogger());
+    jKubeServiceHub = mock(JKubeServiceHub.class, RETURNS_DEEP_STUBS);
+    final OpenShiftClient oc = mock(OpenShiftClient.class);
+    when(jKubeServiceHub.getClient()).thenReturn(oc);
+    when(jKubeServiceHub.getClusterAccess().createDefaultClient()).thenReturn(oc);
+    when(oc.adapt(OpenShiftClient.class)).thenReturn(oc);
     //  @formatter:off
     imageConfiguration = ImageConfiguration.builder()
         .name("foo/bar:latest")
@@ -72,38 +73,38 @@ public class OpenShiftBuildServiceTest {
             .build())
         .build();
     // @formatter:on
-
   }
 
   @Test
   public void push_withEmptyList_shouldNotLogWarning() throws JKubeServiceException {
     // Given
-    when(jKubeServiceHub.getLog()).thenReturn(mockedKitLogger);
+    when(jKubeServiceHub.getLog()).thenReturn(kitLogger);
 
     // When
     new OpenshiftBuildService(jKubeServiceHub).push(Collections.emptyList(), 0, new RegistryConfig(), false);
     // Then
-    verify(mockedKitLogger, times(0)).warn("Image is pushed to OpenShift's internal registry during oc:build goal. Skipping...");
+    verify(kitLogger, times(0)).warn("Image is pushed to OpenShift's internal registry during oc:build goal. Skipping...");
   }
 
   @Test
   public void push_withValidImage_shouldLogWarning() throws JKubeServiceException {
     // Given
-    when(jKubeServiceHub.getLog()).thenReturn(mockedKitLogger);
+    when(jKubeServiceHub.getLog()).thenReturn(kitLogger);
 
     // When
     new OpenshiftBuildService(jKubeServiceHub).push(Collections.singletonList(imageConfiguration), 0, new RegistryConfig(), false);
     // Then
-    verify(mockedKitLogger, times(1)).warn("Image is pushed to OpenShift's internal registry during oc:build goal. Skipping...");
+    verify(kitLogger, times(1)).warn("Image is pushed to OpenShift's internal registry during oc:build goal. Skipping...");
   }
 
   @Test
   public void initClient_withNoOpenShift_shouldThrowException() {
     // Given
+    when(jKubeServiceHub.getClient().adapt(OpenShiftClient.class).isSupported()).thenReturn(false);
     OpenshiftBuildService openshiftBuildService = new OpenshiftBuildService(jKubeServiceHub);
-
-    // When + Then
+    // When
     IllegalStateException illegalStateException = assertThrows(IllegalStateException.class, () -> openshiftBuildService.build(imageConfiguration));
+    // Then
     assertThat(illegalStateException.getMessage())
         .isEqualTo("OpenShift platform has been specified but OpenShift has not been detected!");
   }