+++ kubernetes-maven-plugin/plugin/src/test/java/org/eclipse/jkube/maven/plugin/mojo/develop/LogMojoTest.java
@@ -19,6 +19,7 @@ import java.net.URI;
 import java.util.Properties;
 
 import io.fabric8.kubernetes.client.KubernetesClient;
+import io.fabric8.openshift.client.OpenShiftClient;
 import org.eclipse.jkube.kit.common.KitLogger;
 import org.eclipse.jkube.kit.config.access.ClusterAccess;
 import org.eclipse.jkube.kit.config.service.ApplyService;
@@ -41,6 +42,7 @@ import static org.mockito.ArgumentMatchers.eq;
 import static org.mockito.ArgumentMatchers.isNotNull;
 import static org.mockito.ArgumentMatchers.isNull;
 import static org.mockito.Mockito.RETURNS_DEEP_STUBS;
+import static org.mockito.Mockito.doReturn;
 import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.mockConstruction;
 import static org.mockito.Mockito.times;
@@ -65,7 +67,11 @@ public class LogMojoTest {
   @Before
   public void setUp() throws IOException {
     jKubeServiceHubMockedConstruction = mockConstruction(JKubeServiceHub.class,
-        withSettings().defaultAnswer(RETURNS_DEEP_STUBS));
+        withSettings().defaultAnswer(RETURNS_DEEP_STUBS), (mock, context) -> {
+          final OpenShiftClient oc = mock(OpenShiftClient.class, RETURNS_DEEP_STUBS);
+          doReturn(oc).when(oc).adapt(OpenShiftClient.class);
+          when(mock.getClient()).thenReturn(oc);
+        });
     clusterAccessMockedConstruction = mockConstruction(ClusterAccess.class);
     podLogServiceMockedConstruction = mockConstruction(PodLogService.class);
     kubernetesManifestFile = temporaryFolder.newFile("kubernetes.yml");